refactor(core): clean things up
